requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
iso8601==0.1.13
libscrc==1.5
mypy-extensions==0.4.3
pathspec==0.8.0
pyserial==3.4
regex==2020.10.23
requests==2.24.0
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.25.11
pycrypto==2.6.1
